+++ hat/assets/js/apps/Iaso/domains/registry/components/map/MapTooltip.tsx
